added functionality for the comments.js controller file
